// sdk/storage/storage-file-share/test/utils/constants.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

export enum EnvVarKeys {
  TEST_MODE = "TEST_MODE",
  SUBSCRIPTION_ID = "SUBSCRIPTION_ID",
  RESOURCE_GROUP = "RESOURCE_GROUP",
  ACCOUNT_NAME = "ACCOUNT_NAME",
  ACCOUNT_FILE_URL = "ACCOUNT_FILE_URL",
  ACCOUNT_KEY = "ACCOUNT_KEY",
  ACCOUNT_SAS = "ACCOUNT_SAS",
  STORAGE_CONNECTION_STRING = "STORAGE_CONNECTION_STRING",
  STORAGE_CONNECTION_STRING_WITH_SAS = "STORAGE_CONNECTION_STRING_WITH_SAS",
  SOFT_DELETE_ACCOUNT_NAME = "SOFT_DELETE_ACCOUNT_NAME",
  SOFT_DELETE_ACCOUNT_FILE_URL = "SOFT_DELETE_ACCOUNT_FILE_URL",
  SOFT_DELETE_ACCOUNT_KEY = "SOFT_DELETE_ACCOUNT_KEY",
  SOFT_DELETE_ACCOUNT_SAS = "SOFT_DELETE_ACCOUNT_SAS",
  SOFT_DELETE_STORAGE_CONNECTION_STRING = "SOFT_DELETE_STORAGE_CONNECTION_STRING",
  SOFT_DELETE_STORAGE_CONNECTION_STRING_WITH_SAS = "SOFT_DELETE_STORAGE_CONNECTION_STRING_WITH_SAS",
  PREMIUM_FILE_ACCOUNT_NAME = "PREMIUM_FILE_ACCOUNT_NAME",
  PREMIUM_FILE_ACCOUNT_FILE_URL = "PREMIUM_FILE_ACCOUNT_FILE_URL",
  PREMIUM_FILE_ACCOUNT_KEY = "PREMIUM_FILE_ACCOUNT_KEY",
  PREMIUM_FILE_ACCOUNT_SAS = "PREMIUM_FILE_ACCOUNT_SAS",
  PREMIUM_FILE_STORAGE_CONNECTION_STRING = "PREMIUM_FILE_STORAGE_CONNECTION_STRING",
  PREMIUM_FILE_STORAGE_CONNECTION_STRING_WITH_SAS = "PREMIUM_FILE_STORAGE_CONNECTION_STRING_WITH_SAS",
}

export const ACCOUNT_NAME = "fakeaccountname";
export const ACCOUNT_FILE_URL = `https://${ACCOUNT_NAME}.file.core.windows.net/`;
export const ACCOUNT_KEY = "fakeaccountkey";
export const ACCOUNT_SAS =
  "sv=Sanitized&ss=bfqt&srt=sco&sp=rwdlacup&se=2025-09-03T18%3A13%3A39.8140000Z&sig=sanitizedsig";
export const STORAGE_CONNECTION_STRING = `DefaultEndpointsProtocol=https;AccountName=${ACCOUNT_NAME};AccountKey=${ACCOUNT_KEY};EndpointSuffix=core.windows.net`;
export const STORAGE_CONNECTION_STRING_WITH_SAS = `FileEndpoint=${ACCOUNT_FILE_URL};SharedAccessSignature=${ACCOUNT_SAS}`;
export const SOFT_DELETE_ACCOUNT_NAME = "fakesoftdeleteaccountname";
export const SOFT_DELETE_ACCOUNT_FILE_URL = `https://${SOFT_DELETE_ACCOUNT_NAME}.file.core.windows.net/`;
export const SOFT_DELETE_ACCOUNT_KEY = "fakesoftdeleteaccountkey";
export const SOFT_DELETE_ACCOUNT_SAS =
  "sv=Sanitized&ss=bfqt&srt=sco&sp=rwdlacup&se=2025-09-03T18%3A13%3A39.8140000Z&sig=sanitizedsig";
export const SOFT_DELETE_STORAGE_CONNECTION_STRING = `DefaultEndpointsProtocol=https;AccountName=${SOFT_DELETE_ACCOUNT_NAME};AccountKey=${SOFT_DELETE_ACCOUNT_KEY};EndpointSuffix=core.windows.net`;
export const SOFT_DELETE_STORAGE_CONNECTION_STRING_WITH_SAS = `FileEndpoint=${SOFT_DELETE_ACCOUNT_FILE_URL};SharedAccessSignature=${SOFT_DELETE_ACCOUNT_SAS}`;
export const PREMIUM_FILE_ACCOUNT_NAME = "fakepremiumfileaccountname";
export const PREMIUM_FILE_ACCOUNT_FILE_URL = `https://${PREMIUM_FILE_ACCOUNT_NAME}.file.core.windows.net/`;
export const PREMIUM_FILE_ACCOUNT_KEY = "fakepremiumfileaccountkey";
export const PREMIUM_FILE_ACCOUNT_SAS =
  "sv=Sanitized&ss=bfqt&srt=sco&sp=rwdlacup&se=2025-09-03T18%3A13%3A39.8140000Z&sig=sanitizedsig";
export const PREMIUM_FILE_STORAGE_CONNECTION_STRING = `DefaultEndpointsProtocol=https;AccountName=${PREMIUM_FILE_ACCOUNT_NAME};AccountKey=${PREMIUM_FILE_ACCOUNT_KEY};EndpointSuffix=core.windows.net`;
export const PREMIUM_FILE_STORAGE_CONNECTION_STRING_WITH_SAS = `FileEndpoint=${PREMIUM_FILE_ACCOUNT_FILE_URL};SharedAccessSignature=${PREMIUM_FILE_ACCOUNT_SAS}`;